test(report): split viewer renderer contracts
